delete WeatherApp.tsx -> App.tsx
